backend/requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
psycopg2-binary==2.9.7
Pygments==2.16.1
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
